Rename the patterns source to represent this repo
